refactor: update link and toggle button filenames
