+++ design-toolkit/components/src/components/action-bar/action-bar.stories.tsx
@@ -1,6 +1,6 @@
 import { Placeholder } from '@accelint/icons';
 import { Button } from '../button';
-import { ToggleButton } from '../button/toggle-button';
+import { ToggleButton } from '../button/toggle';
 import { Icon } from '../icon';
 import { ActionBar } from './index';
 /*
