+++ design-toolkit/components/src/components/button/button.stories.tsx
@@ -14,8 +14,8 @@ import { Placeholder } from '@accelint/icons';
 import { composeRenderProps } from 'react-aria-components';
 import { Icon } from '../icon';
 import { Button } from './';
-import { LinkButton } from './link-button';
-import { ToggleButton } from './toggle-button';
+import { LinkButton } from './link';
+import { ToggleButton } from './toggle';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta = {
