+++ design-toolkit/components/src/index.ts
@@ -58,14 +58,14 @@ export {
   ToggleButtonProvider,
 } from './components/button/context';
 export { Button } from './components/button';
-export { LinkButton } from './components/button/link-button';
+export { LinkButton } from './components/button/link';
 export {
   ButtonStyles,
   ButtonStylesDefaults,
   LinkButtonStyles,
   ToggleButtonStyles,
 } from './components/button/styles';
-export { ToggleButton } from './components/button/toggle-button';
+export { ToggleButton } from './components/button/toggle';
 export type {
   ButtonProps,
   ButtonStyleVariants,
